rename d500 emitter test, cover emitter always on and mutual exclusion
